+++ workspaces/service/src/service/api/ContractFactory.ts
@@ -11,6 +11,7 @@ import { ApiPromise } from '@polkadot/api';
 import { Abi, ContractPromise } from '@polkadot/api-contract';
 import { KeyringPair } from '@polkadot/keyring/types';
 import type { ContractInstantiateResult } from '@polkadot/types/interfaces/contracts';
+import { BN } from '@polkadot/util/bn';
 
 
 export type CreateOptions = {
@@ -122,10 +123,12 @@ export class ContractFactory<T extends Contract = Contract>
             ...options
         };
         
-        const keyringPair : any = typeof options.asAccount === 'string'
-            ? this._devPhase.accounts[options.asAccount]
-            : options.asAccount;
-            
+        const keyringPair : KeyringPair = <any> (
+            typeof options.asAccount === 'string'
+                ? this._devPhase.accounts[options.asAccount]
+                : options.asAccount
+        );
+        
         // verify is it required to upload resource
         if (!options.forceUpload) {
             if (!this._systemContract) {
@@ -135,10 +138,7 @@ export class ContractFactory<T extends Contract = Contract>
                 );
             }
         
-            const cert = await PhalaSdk.signCertificate({
-                api: this.api,
-                pair: keyringPair
-            });
+            const cert = await PhalaSdk.signCertificate({ pair: keyringPair });
             
             const codeType = ContractFactory.CODE_TYPE_MAP[this.contractType];
             if (!codeType) {
@@ -149,8 +149,8 @@ export class ContractFactory<T extends Contract = Contract>
             }
             
             const { output } = await this._systemContract.query['system::codeExists'](
-                cert,
-                {},
+                keyringPair.address,
+                { cert },
                 this.metadata.source.hash,
                 codeType
             );
@@ -285,23 +285,22 @@ export class ContractFactory<T extends Contract = Contract>
     ) : Promise<ContractInstantiateResult>
     {
         const systemContract = await this._devPhase.getSystemContract(this.clusterId);
-        
+
         const abi = new Abi(this.metadata);
         const callData = abi.findConstructor(constructor).toU8a(params);
         const salt = typeof options.salt == 'number'
             ? '0x' + options.salt.toString(16)
             : options.salt
         ;
-        
+
         const keyringPair : any = typeof options.asAccount === 'string'
             ? this._devPhase.accounts[options.asAccount]
             : options.asAccount;
-        
-        const cert = await PhalaSdk.signCertificate({
-            api: this.api,
-            pair: keyringPair
-        });
-        
+
+        const cert = await PhalaSdk.signCertificate({ pair: keyringPair });
+
+        // todo ld 2023-08-10 01:16:30
+        // @ts-ignore
         const instantiateReturn = await systemContract.instantiate({
             codeHash: this.metadata.source.hash,
             salt,
@@ -310,6 +309,8 @@ export class ContractFactory<T extends Contract = Contract>
             transfer: options.transfer
         }, cert);
         
+        console.log(instantiateReturn);
+
         return <any> instantiateReturn;
     }
     
@@ -320,24 +321,50 @@ export class ContractFactory<T extends Contract = Contract>
     {
         const api : any = await this._devPhase.createApiPromise();
         
-        const phala : any = await PhalaSdk.create({
-            api: api,
-            baseURL: this._devPhase.workerUrl,
-            contractId,
-            autoDeposit: true,
-        });
+        const clusterInfo = (
+            await api.query.phalaPhatContracts.clusters(this.clusterId)
+        ).toJSON();
         
-        const instance = new ContractPromise(
-            phala.api,
+        const onChainRegistry = await PhalaSdk.OnChainRegistry.create(
+            api,
+            {
+                clusterId: this.clusterId,
+                pruntimeURL: this._devPhase.workerUrl,
+                workerId: this._devPhase.workerInfo.publicKey,
+                systemContractId: this._systemContract?.contractId,
+                autoConnect: true,
+                skipCheck: true, // todo ld 2023-08-10 23:45:02 - dirty hack!
+            }
+        );
+        // todo ld 2023-08-15 06:07:03 - dirty hack!
+        onChainRegistry.clusterInfo = {
+            ...clusterInfo,
+            gasPrice: new BN(1)
+        };
+        
+        const contractKey = await onChainRegistry.getContractKey(contractId);
+        if (!contractKey) {
+            throw new Exception(
+                'Contract key is not ready yet',
+                1691622666843
+            );
+        }
+        
+        const instance = new PhalaSdk.PinkContractPromise(
+            api,
+            onChainRegistry,
             this.metadata,
             contractId,
+            contractKey
         );
         
         Object.assign(instance, {
             contractId,
             clusterId: this.clusterId,
-            sidevmQuery: phala.sidevmQuery,
-            instantiate: phala.instantiate,
+            
+            // todo ld 2023-08-10 01:11:22
+            // sidevmQuery: phala.sidevmQuery,
+            // instantiate: phala.instantiate,
         });
         
         return <any>instance;