+++ workspaces/xsandbox/tests/flipper/fliper.test.ts
@@ -8,14 +8,11 @@ describe('Flipper', () => {
     let factory : Flipper.Factory;
     let contract : Flipper.Contract;
     let signer : KeyringPair;
-    let certificate : PhalaSdk.CertificateData;
+    let cert : PhalaSdk.CertificateData;
     
     before(async function() {
         signer = this.devPhase.accounts.bob;
-        certificate = await PhalaSdk.signCertificate({
-            api: this.api,
-            pair: signer,
-        });
+        cert = await PhalaSdk.signCertificate({ pair: signer });
     });
     
     beforeEach(async function() {
@@ -33,7 +30,7 @@ describe('Flipper', () => {
         });
         
         it('Should be created with proper intial value', async function() {
-            const response = await contract.query.get(certificate, {});
+            const response = await contract.query.get(signer.address, { cert });
             expect(response.output.toJSON()).to.be.eql({ ok: false });
         });
     });
@@ -44,7 +41,7 @@ describe('Flipper', () => {
         });
         
         it('Should be created with proper intial value', async function() {
-            const response = await contract.query.get(certificate, {});
+            const response = await contract.query.get(signer.address, { cert });
             expect(response.output.toJSON()).to.be.eql({ ok: true });
         });
     });