+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v0.6.0 / 2023-08-15
+- breaking changes in ContractFactory, contract interfaces etc.
+- reflect changes in latest @phala/sdk
+
 ### v0.5.0 / 2023-07-25
 - next big improvement of type binding
 - dependencies update